--- bundles/liberty/src/liberty/tools/LibertyNature.java
@@ -1,44 +1,0 @@
-/*******************************************************************************
-* Copyright (c) 2022 IBM Corporation and others.
-*
-* This program and the accompanying materials are made available under the
-* terms of the Eclipse Public License v. 2.0 which is available at
-* http://www.eclipse.org/legal/epl-2.0.
-*
-* SPDX-License-Identifier: EPL-2.0
-*
-* Contributors:
-*     IBM Corporation - initial implementation
-*******************************************************************************/
-package liberty.tools;
-
-import org.eclipse.core.resources.IProject;
-import org.eclipse.core.resources.IProjectNature;
-import org.eclipse.core.runtime.CoreException;
-
-/**
- * Represents a Liberty nature or type.
- */
-public class LibertyNature implements IProjectNature {
-    // NATURE_ID must follow the following convention: <bundle-symbolic-name> . <nature-id-as-per-plugin.xml>
-    public static final String NATURE_ID = "liberty.libertyNature";
-    private IProject project;
-
-    @Override
-    public void configure() throws CoreException {
-    }
-
-    @Override
-    public void deconfigure() throws CoreException {
-    }
-
-    @Override
-    public IProject getProject() {
-        return project;
-    }
-
-    @Override
-    public void setProject(IProject project) {
-        this.project = project;
-    }
-}
