+++ tests/META-INF/MANIFEST.MF
@@ -8,8 +8,8 @@ Bundle-RequiredExecutionEnvironment: JavaSE-11
 Require-Bundle: org.junit.jupiter.api;bundle-version="[5.8.0,6.0.0)",
  org.eclipse.ui,
  org.hamcrest.library;bundle-version="[1.3.0,2.2)"
-Import-Package: liberty.tools,
- liberty.tools.ui,
+Import-Package: io.openliberty.tools.eclipse,
+ io.openliberty.tools.eclipse.ui,
  org.eclipse.core.resources,
  org.eclipse.core.runtime,
  org.eclipse.m2e.core,
