+++ tests/src/io/openliberty/tools/eclipse/test/it/LibertyPluginSWTBotDashboardTest.java
@@ -1,0 +1,74 @@
+/*******************************************************************************
+* Copyright (c) 2022 IBM Corporation and others.
+*
+* This program and the accompanying materials are made available under the
+* terms of the Eclipse Public License v. 2.0 which is available at
+* http://www.eclipse.org/legal/epl-2.0.
+*
+* SPDX-License-Identifier: EPL-2.0
+*
+* Contributors:
+*     IBM Corporation - initial implementation
+*******************************************************************************/
+package io.openliberty.tools.eclipse.test.it;
+
+import org.eclipse.swtbot.eclipse.finder.SWTWorkbenchBot;
+import org.eclipse.swtbot.eclipse.finder.widgets.SWTBotView;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+
+import io.openliberty.tools.eclipse.test.it.utils.SWTPluginOperations;
+
+/**
+ * Tests Open Liberty Eclipse plugin functions.
+ */
+public class LibertyPluginSWTBotDashboardTest {
+
+    /**
+     * Wokbench bot instance.
+     */
+    static SWTWorkbenchBot bot;
+
+    /**
+     * Dashboard instance.
+     */
+    static SWTBotView dashboard;
+    
+    /**
+     * Setup.
+     */
+    @BeforeAll
+    public static void setup() {
+        bot = new SWTWorkbenchBot();
+        SWTPluginOperations.closeWelcomePage(bot);
+    }
+
+    /**
+     * Cleanup.
+     */
+    @AfterAll
+    public static void cleanup() {
+        bot.closeAllEditors();
+        bot.closeAllShells();
+        bot.resetWorkbench();
+    }
+
+    /**
+     * Tests opening the dashboard using the main toolbar icon.
+     */
+    @Test
+    public void testOpenDashboardWithToolbarIcon() {
+        // Open the dashboard view.
+        SWTPluginOperations.openDashboardUsingToolbar(bot);
+    }
+
+    /**
+     * Tests opening the dashboard using the Liberty menu.
+     */
+    @Test
+    public void testOpenDashboardUsingMenu() {
+        // Open the dashboard view.
+        SWTPluginOperations.openDashboardUsingMenu(bot);
+    }
+}
